+++ resources/firewalld.rb
@@ -3,11 +3,15 @@ unified_mode true
 provides :firewalld,
          os: 'linux'
 
+# Custom options to pass to the package manager during install of firewalld package
+property :package_options, String
+
 action :install do
   chef_gem 'ruby-dbus'
   require 'dbus'
 
   package 'firewalld' do
+    options new_resource.package_options if new_resource.package_options
     action :install
   end