+++ test/fixtures/cookbooks/firewall-test/recipes/default.rb
@@ -1,3 +1,8 @@
+
+firewalld = rhel? || amazon_linux?
+iptables = !firewalld && node['firewall']['ubuntu_iptables']
+ufw = !firewalld && !iptables
+
 include_recipe 'firewall'
 
 firewall_rule 'ssh22' do
@@ -35,7 +40,7 @@ end
 firewall_rule 'protocolnum' do
   protocol 112
   command :allow
-  only_if { rhel? || amazon_linux? || node['firewall']['ubuntu_iptables'] } # debian ufw doesn't support protocol numbers
+  not_if { ufw } # debian ufw doesn't support protocol numbers
 end
 
 firewall_rule 'prepend' do
@@ -59,52 +64,52 @@ end
 firewall_rule 'range' do
   port 1000..1100
   command :allow
-
-  # centos 5 is broken for ipv6 ranges
-  # see https://github.com/chef-cookbooks/firewall/pull/111#issuecomment-163520156
-  not_if { rhel? && node['platform_version'].to_f < 6.0 }
 end
 
 firewall_rule 'array' do
   port [1234, 5000..5100, '5678']
   command :allow
-
-  # centos 5 is broken for ipv6 ranges
-  # see https://github.com/chef-cookbooks/firewall/pull/111#issuecomment-163520156
-  not_if { rhel? && node['platform_version'].to_f < 6.0 }
 end
 
-# if using with iptables-restart, this produces an unreadable line; no problem, IF disabled
-firewall_rule 'ufw raw test' do
-  raw 'allow from 192.168.1.1 to 192.168.2.1 port 25 proto tcp'
-  only_if { platform_family?('debian') && !node['firewall']['ubuntu_iptables'] }
+if firewalld
+  firewall_rule 'allow 5672 in internal zone' do
+    zone 'internal'
+    port 5672
+  end
 end
 
-firewall_rule 'RPC Port Range In' do
-  port 5000..5100
-  protocol :tcp
-  command :allow
-  direction :in
-
-  # centos 5 is broken for ipv6 ranges
-  # see https://github.com/chef-cookbooks/firewall/pull/111#issuecomment-163520156
-  not_if { rhel? && node['platform_version'].to_f < 6.0 }
+if ufw
+  firewall_rule 'ufw raw test' do
+    raw 'allow from 192.168.1.1 to 192.168.2.1 port 25 proto tcp'
+  end
 end
 
-firewall_rule 'HTTP HTTPS' do
-  port [80, 443]
-  protocol :tcp
-  direction :out
-  command :allow
-end
+if iptables
+  firewall_rule 'RPC Port Range In' do
+    port 5000..5100
+    protocol :tcp
+    command :allow
+    direction :in
 
-firewall_rule 'port2433' do
-  description 'This should not be included'
-  include_comment false
-  source    '127.0.0.0/8'
-  port      2433
-  direction :in
-  command   :allow
+    # # centos 5 is broken for ipv6 ranges
+    # # see https://github.com/chef-cookbooks/firewall/pull/111#issuecomment-163520156
+    # not_if { rhel? && node['platform_version'].to_f < 6.0 }
+  end
+  firewall_rule 'HTTP HTTPS' do
+    port [80, 443]
+    protocol :tcp
+    direction :out
+    command :allow
+  end
+
+  firewall_rule 'port2433' do
+    description 'This should not be included'
+    include_comment false
+    source    '127.0.0.0/8'
+    port      2433
+    direction :in
+    command   :allow
+  end
 end
 
 include_recipe 'firewall-test::windows' if windows?